find by cin

+++ src/main/java/com/example/tnb/repository/TauxRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 
 public interface TauxRepository extends JpaRepository<Taux, Integer> {
-    Taux findById(int id);
+    Taux findById(long id);
 }
